+++ popelnukhProj/src/test/java/providersTest/AddNewProviderTest.java
@@ -1,6 +1,5 @@
 package providersTest;
 
-import LoginTest.LoginWithPageObjectTest;
 import abstractParentTest.AbstractParentTest;
 import org.junit.Test;
