+++ vetrichenkoProj/src/main/java/endpoints/EndPoints.java
@@ -1,0 +1,7 @@
+package endpoints;
+
+public interface EndPoints {
+	String PRIVATBANK_P24API_PUBINFO="http://api.privatbank.ua/p24api/pubinfo";
+
+	String RESTAPI_DEMOQA_COM_CUSTOMER_REGISTER="http://restapi.demoqa.com/customer/register";
+}
