+++ vetrichenkoProj/src/test/java/abstractParentTest/AbstractParentTest.java
@@ -1,6 +1,8 @@
 package abstractParentTest;
 
 import io.github.bonigarcia.wdm.WebDriverManager;
+import libs.ConfigProperties;
+import org.aeonbits.owner.ConfigFactory;
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
@@ -21,6 +23,9 @@ public class AbstractParentTest {
     protected HomePage homePage;
     protected SparesPage sparePage;
     protected EditSparesPage editSparesPage;
+
+    protected static ConfigProperties configProperties =
+            ConfigFactory.create(ConfigProperties.class);
 
     @Before
     public void setUp() throws Exception {
