+++ vetrichenkoProj/src/test/java/suits/SmokeSuit.java
@@ -1,6 +1,6 @@
 package suits;
 
-import LoginTest.LoginWithPageObjectTest;
+import loginTest.LoginWithPageObjectTest;
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 import sparesTest.SparesTest;
